pyproject.toml
[project]
name = "hellosubagents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-ollama>=1.0.0",
    "langgraph-cli[inmem]>=0.4.19",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "agentevals>=0.0.9",
    "pytest>=9.0.2",
]
